backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.9
pandas>=2.2.0
aiofiles>=23.2.1
